feat: Integrate Plyr video player
Integrates the Plyr.io library to provide a more robust and customizable video player experience. This includes importing the necessary CSS and JavaScript for the player, as well as refactoring the `VideoPlayer` component to utilize Plyr for YouTube embeds. The `getCleanId` function has also been enhanced to better handle various input formats for video IDs.